requirements.txt
lightkurve>=2.0.1
numpy
scipy
matplotlib
batman-package
emcee
astroquery
tqdm
